split logic from dom
