+++ src/functions/functions.js
@@ -1,5 +1,38 @@
 import NewList from "./newList";
 import NewTask from "./newTask";
+import events from "../functions/eventListeners";
+import { isToday } from "date-fns";
+import createTaskElement from "../dom/createTaskElement";
+
+export function updateTasks(listName, listArray) {
+    const title = document.querySelector(".title");
+    const container = document.querySelector(".container");
+    container.innerHTML = "";
+    //title
+    title.textContent = listName;
+
+    //tasks
+    listArray.forEach((element) => {
+        if (listName === "Tasks") {
+            element.task.forEach((item) => {
+                createTaskElement(item);
+            });
+        } else if (listName === "My Day") {
+            element.task.forEach((item) => {
+                const dateArray = item.date.split("-");
+                const newDate = `${dateArray[2]}, ${dateArray[1]}, ${dateArray[0]}`;
+                if (isToday(new Date(newDate))) {
+                    createTaskElement(item);
+                }
+            });
+        } else if (element.name === listName) {
+            element.task.forEach((item) => {
+                createTaskElement(item);
+            });
+        }
+    });
+    events().checkBox(listArray);
+}
 
 export function updateLists(listsArray) {
     const lists = document.querySelector(".lists");
