+++ src/index.js
@@ -1,8 +1,7 @@
 import Content from "./dom/tasksDOM";
 import events from "./functions/eventListeners";
 import Sidebar from "./dom/sidebar";
-import { updateLists, getLocalStorage } from "./functions/functions";
-import updateMain from "./dom/updateTasksDOM.js";
+import { updateTasks, updateLists, getLocalStorage } from "./functions/functions";
 import "./styles/style.css";
 
 let listsArray = [];
@@ -11,7 +10,7 @@ getLocalStorage(listsArray);
 Sidebar();
 Content();
 const myDay = document.querySelector(".my-day");
-updateMain(myDay.textContent, listsArray);
+updateTasks(myDay.textContent, listsArray);
 updateLists(listsArray);
 events().sidebarLists(listsArray);
 events().newList(listsArray);
